// src/features/assessments/types/assessment.types.ts
export type AssessmentListItem = {
  id: string;
  assessmentName: string;
  description: string;
  educationStage: string;
  grade: string;
  subject: string;
  questionType: string;
  totalQuestions: number;
  totalMarks: number;
  durationMinutes: number;
  difficultyLevel: string;
  assessmentDate: string;
  status: string;
  createdDate: string;
};